refactor: 폴더 구조 단순화
shadcn ui 설치시 받았던 utils.ts를 utils폴더 안에 넣었습니다.
